cockblocks all over again

+++ code/modules/roguetown/roguemachine/antagcommentor.dm
@@ -1,6 +1,20 @@
 /datum/mind
 	var/freehold_token_touched = FALSE // I need to test it before mob/living/carbon/human because druids shitters going to abuse shit out of their ass broken spell
 
+
+/obj/structure/roguemachine/freeholdpress/proc/target_apply_normal_reward(mob/living/carbon/human/H, reward_type) //helper help me
+	if(!H || !reward_type)
+		return
+	if(reward_type == /datum/status_effect/buff/freehold_normal_faith_spdper)
+		H.apply_status_effect(/datum/status_effect/buff/freehold_normal_faith_spdper)
+		return
+	if(reward_type == /datum/status_effect/buff/freehold_normal_faith_strcon)
+		H.apply_status_effect(/datum/status_effect/buff/freehold_normal_faith_strcon)
+		return
+	if(reward_type == /datum/status_effect/buff/freehold_normal_faith_intwil)
+		H.apply_status_effect(/datum/status_effect/buff/freehold_normal_faith_intwil)
+		return
+
 /obj/structure/roguemachine/freeholdpress
 	name = "Karmic eye"
 	desc = "A machine that turns your karma into a reward."
@@ -13,57 +27,93 @@
 /obj/structure/roguemachine/freeholdpress/attack_hand(mob/user)
 	if(!ishuman(user))
 		return
-
-	var/mob/living/carbon/human/H = user
-	
-	if(HAS_TRAIT(H, TRAIT_FREEHOLDER))
-		if(!H.has_status_effect(/datum/status_effect/buff/freehold_great_faith))
-			to_chat(H, span_warning("Im supposed to be marked by another person to claim anything from here."))
-			playsound(src, 'sound/misc/machineno.ogg', 100, FALSE, -1)
-			return
-
-		var/list/choices = list(
-			"Strength",
-			"Constitution",
-			"Speed",
-			"Perception",
-			"Intelligence",
-			"Willpower"
-		)
-
-		var/choice = input(H, "Choose your reward to claim.", src.name) as null|anything in choices
-		if(!choice)
-			return
-
-		switch(choice)
-			if("Strength")
-				H.STASTR += 1
-			if("Constitution")
-				H.STACON += 1
-			if("Speed")
-				H.STASPD += 1
-			if("Perception")
-				H.STAPER += 1
-			if("Intelligence")
-				H.STAINT += 1
-			if("Willpower")
-				H.STAWIL += 1
-
-		H.remove_status_effect(/datum/status_effect/buff/freehold_great_faith)
-		to_chat(H, span_notice("I'm accepting my reward from the machine."))
-		playsound(src, 'sound/misc/machinetalk.ogg', 100, FALSE, -1)
+	if(!Adjacent(user))
 		return
 
+	var/mob/living/carbon/human/H = user
+
+	if(HAS_TRAIT(H, TRAIT_FREEHOLDER))
+		if(H.has_status_effect(/datum/status_effect/buff/freehold_great_faith))
+			var/list/great_choices = list(
+				"Strength",
+				"Constitution",
+				"Speed",
+				"Perception",
+				"Intelligence",
+				"Willpower"
+			)
+
+			var/great_choice = input(H, "Choose your permanent reward to claim.", src.name) as null|anything in great_choices
+			if(!great_choice)
+				return
+			if(QDELETED(src) || QDELETED(H))
+				return
+			if(!Adjacent(H))
+				return
+			if(!H.has_status_effect(/datum/status_effect/buff/freehold_great_faith))
+				to_chat(H, span_warning("The greater blessing has already faded."))
+				playsound(src, 'sound/misc/machineno.ogg', 100, FALSE, -1)
+				return
+
+			switch(great_choice)
+				if("Strength")
+					H.STASTR += 1
+				if("Constitution")
+					H.STACON += 1
+				if("Speed")
+					H.STASPD += 1
+				if("Perception")
+					H.STAPER += 1
+				if("Intelligence")
+					H.STAINT += 1
+				if("Willpower")
+					H.STAWIL += 1
+
+			H.remove_status_effect(/datum/status_effect/buff/freehold_great_faith)
+			to_chat(H, span_notice("I'm accepting a great reward from the machine."))
+			playsound(src, 'sound/misc/machinetalk.ogg', 100, FALSE, -1)
+			return
+
+		if(H.has_status_effect(/datum/status_effect/buff/freehold_normal_faith))
+			var/list/normal_choices = list(
+				"Speed and Perception",
+				"Strength and Constitution",
+				"Intelligence and Willpower"
+			)
+
+			var/normal_choice = input(H, "Choose your temporary reward to claim.", src.name) as null|anything in normal_choices
+			if(!normal_choice)
+				return
+			if(QDELETED(src) || QDELETED(H))
+				return
+			if(!Adjacent(H))
+				return
+			if(!H.has_status_effect(/datum/status_effect/buff/freehold_normal_faith))
+				to_chat(H, span_warning("The blessing has already faded."))
+				playsound(src, 'sound/misc/machineno.ogg', 100, FALSE, -1)
+				return
+			switch(normal_choice)
+				if("Speed and Perception")
+					target_apply_normal_reward(H, /datum/status_effect/buff/freehold_normal_faith_spdper)
+				if("Strength and Constitution")
+					target_apply_normal_reward(H, /datum/status_effect/buff/freehold_normal_faith_strcon)
+				if("Intelligence and Willpower")
+					target_apply_normal_reward(H, /datum/status_effect/buff/freehold_normal_faith_intwil)
+			H.remove_status_effect(/datum/status_effect/buff/freehold_normal_faith)
+			to_chat(H, span_notice("I'm accepting a lesser reward from the machine."))
+			playsound(src, 'sound/misc/machinetalk.ogg', 100, FALSE, -1)
+			return
+		to_chat(H, span_warning("Im supposed to be marked by another person to claim anything from here."))
+		playsound(src, 'sound/misc/machineno.ogg', 100, FALSE, -1)
+		return
 	if(H.mind?.freehold_token_touched)
 		to_chat(H, span_warning("This machine will grant me no further token. Only once per my life."))
 		playsound(src, 'sound/misc/machineno.ogg', 100, FALSE, -1)
 		return
-
 	for(var/obj/item/freehold_token/T in H.contents)
 		to_chat(H, span_warning("I already carry such a token."))
 		playsound(src, 'sound/misc/machineno.ogg', 100, FALSE, -1)
 		return
-
 	if(H.mind)
 		H.mind.freehold_token_touched = TRUE
 	to_chat(H, span_notice("The machine presses a fleeting token into my hand."))
@@ -198,12 +248,12 @@
 			apply_great_faith(user, target)
 
 /obj/item/freehold_token/proc/apply_bad_faith(mob/living/carbon/human/user, mob/living/carbon/human/target)
+	target.remove_status_effect(/datum/status_effect/buff/freehold_normal_faith)
+	target.remove_status_effect(/datum/status_effect/buff/freehold_great_faith)
 	target.remove_status_effect(/datum/status_effect/buff/freehold_normal_faith_spdper)
 	target.remove_status_effect(/datum/status_effect/buff/freehold_normal_faith_strcon)
 	target.remove_status_effect(/datum/status_effect/buff/freehold_normal_faith_intwil)
-
 	target.apply_status_effect(/datum/status_effect/debuff/freehold_bad_faith)
-
 	user.visible_message(span_warning("[user] lays a token of bad faith upon [target]."))
 	to_chat(target, span_warning("You have been marked by bad faith."))
 	playsound(target, 'sound/misc/machineno.ogg', 100, FALSE, -1)
@@ -211,15 +261,18 @@
 
 /obj/item/freehold_token/proc/apply_normal_faith(mob/living/carbon/human/user, mob/living/carbon/human/target)
 	if(target.has_status_effect(/datum/status_effect/debuff/freehold_bad_faith))
-		to_chat(user, span_warning("Coal aura still stucks to to [target]."))
+		to_chat(user, span_warning("Coal aura still stucks to [target]."))
 		in_use = FALSE
 		return
-
-	var/list/choices = list(
-		"Speed and Perception",
-		"Strength and Constitution",
-		"Intelligence and Willpower"
-	)
+	if(target.has_status_effect(/datum/status_effect/buff/freehold_normal_faith) || target.has_status_effect(/datum/status_effect/buff/freehold_great_faith))
+		to_chat(user, span_warning("[target] already bears a blessing."))
+		in_use = FALSE
+		return
+	target.apply_status_effect(/datum/status_effect/buff/freehold_normal_faith)
+	user.visible_message(span_notice("[user] lays a token of plain faith upon [target]."))
+	to_chat(target, span_notice("A lesser blessing settles upon you. I should claim it at the machine."))
+	playsound(target, 'sound/misc/machinetalk.ogg', 100, FALSE, -1)
+	qdel(src)
 
 	var/choice = input(user, "Choose the normal level reward (a temporary buff).", src.name) as null|anything in choices
 	if(!choice)
@@ -261,10 +314,13 @@
 		playsound(target, 'sound/misc/machineyes.ogg', 100, FALSE, -1)
 		qdel(src)
 		return
-
+	if(target.has_status_effect(/datum/status_effect/buff/freehold_normal_faith) || target.has_status_effect(/datum/status_effect/buff/freehold_great_faith))
+		to_chat(user, span_warning("[target] already bears a blessing."))
+		in_use = FALSE
+		return
 	target.apply_status_effect(/datum/status_effect/buff/freehold_great_faith)
 	user.visible_message(span_notice("[user] lays a token of great faith upon [target]."))
-	to_chat(target, span_notice("The person has commented my efforts as great. I must claim it quickly at the machine."))
+	to_chat(target, span_notice("My efforts have been greatly praised. I must claim it quickly at the machine."))
 	playsound(target, 'sound/misc/machineyes.ogg', 100, FALSE, -1)
 	qdel(src)
 
@@ -326,9 +382,24 @@
 	alert_type = /atom/movable/screen/alert/status_effect/buff/freehold_great_faith
 	duration = 3 MINUTES
 	status_type = STATUS_EFFECT_REFRESH
-	examine_text = "SUBJECTPRONOUN bears a fleeting promise of greater faith."
 
 /datum/status_effect/buff/freehold_great_faith/on_apply()
 	. = ..()
 	if(ishuman(owner))
 		to_chat(owner, span_notice("I have but a short while to claim a great reward from the machine."))
+
+/atom/movable/screen/alert/status_effect/buff/freehold_normal_faith
+	name = "Normal Faith"
+	desc = "My efforts have been noticed."
+	icon_state = "buff"
+
+/datum/status_effect/buff/freehold_normal_faith
+	id = "freehold_normal_faith"
+	alert_type = /atom/movable/screen/alert/status_effect/buff/freehold_normal_faith
+	duration = 3 MINUTES
+	status_type = STATUS_EFFECT_REFRESH
+
+/datum/status_effect/buff/freehold_normal_faith/on_apply()
+	. = ..()
+	if(ishuman(owner))
+		to_chat(owner, span_notice("I have but a short while to claim a lesser reward from the machine."))
